FORM SUPPORT! IT HAS ARRIVED!!
Code-cleanup: Added 'form' support for input-fields. This has been in the works for nearly a entire month. That said, most of that time was spent crying in a corner with a notepad and a pen next to me.
    > This allows you to connect multiple input-fields into a single 'form'. Kinda like how when you fill out a paper it includes Name, Surname, Address, etc all in a single form.
Code-cleanup: Input-fields now actually have a value after they've been made (previously the value would just be undefined until the user changed the value to something else)
    > Because of this, radio fields now select the first option on default.
Code-cleanup: Added a 'defaultValue' attribute that specifies what value the input-field should have after being created. If the specified 'defaultValue' is invalid it will be ignored.
Code-cleanup: A lot of the code in 'inputfield_createField()' has been moved to 'inputfield_FieldObject()' to avoid repeats.
    > The two functions serve a different purpose now. 'inputfield_createField()' only creates the HTML side of the input-field while 'inputfield_FieldObject()' does basically everything else.
Code-cleanup/Fixed: Input-fields now run completely over HTML 'onchange'/'onclick' attributes instead of manually recreating it using event-listeners. This fixes input-fields not updating when changed through keyboard only and generally behaving weird sometimes.
Code-cleanup/Fixed: Something about hosts not working as supposed to. Though I honestly forgot what the bug actually was because the fix was made like an eternity ago (even though it was yesterday).

Reading this change-log feels weird when you remember this is a Mario Party related project and not some library for accepting user-input.