Input-field cleanup
Code-cleanup: The '.isDOMElement()' function now accepts a optional argument so it can be called using 'Object.isDOMElement(*variable*)'.
Code-cleanup: 'inputfield_getElement()' now supports picking the DOM tree ('document'/DocumentFragments) and using a element for reference (basically a way to speedup the function).
Code-cleanup: Cleaned up the documentation at the beginning of the file.
Code-cleanup: Renamed the 'data-fieldid' to 'data-linktofieldid' and 'data-fieldvalue' to 'data-holdsvalue'
Code-cleanup: Added a test in settings that tests whether using a tag name multiple times breaks input-fields (they did).
Code-cleanup: Fixed not fetching the input-field specified in a 'addToForm' attribute when only the ID is specified. This does not fix when used alongside DocumentFragments. I tried, trust me.
Code-cleanup: Added a bit to the documentation that says that you shouldn't use DocumentFragments in extremely specific cases. Yes, this a bug that is now a feature. Not a good feature but how many features are actually good anyway?
Code-cleanup: Added a error message for the above.