Various improvements and fixes
Fixed: Counters not working. Whoops.
Code-cleanup: Fixed the boot-process documentation at the beginning of 'core/boot.js'
Code-cleanup: Made 'helpers.js' completely self-contained (I guess this is more of a fix since it was a oversight -- this also changes the boot-process slightly).
Code-cleanup: Improved two functions that were badly made and nearly identical (combined 'fillInObject()' and 'forceFillInObject()', renamed it to '.fillIn()' and defined it on the Object prototype).
Code-cleanup: Improved code by re-using code more (created a '.removeEachIf()' function that iterates on an array and removes items as it goes -- this avoids creating a unique backwards for-loop several each time array items have to be removed).