Add controls
Added: Action and amount buttons in the navbar at the top. You can now select whether you want to add or subtract and also by how much.
Added: The ability to set a counter to a certain number. That's the = symbol. This deserves a own line since this is new to the rewrite and didn't exist in the original MPO.
Fixed: A typo in the credits/license text (

Code-cleanup: Made naming more consistent ('value' > 'amount' inside 'trackerCore_status.controls' -- 'updatesTracker_changeCharacter()' > 'updatesTracker_setCharacter()').
Code-cleanup: Added a small documentation at the beginning of 'core/updatesTracker.js'.
Code-cleanup: Added functions to get the default action and amount and used them where needed ('updatesTracker_getAction()' and 'updatesTracker_getAmount()').
Code-cleanup: Made the creation of modals more consistent with other parts of the code (the 'specifics' argument has been split into several own arguments since there's already a catch-all 'attributes' argument - this is now consistent with how input-fields are made).
Code-cleanup: Input-field tags are now always a string (this was a bug that happened when the ID is converted to a tag when no tag has been specified).
Code-cleanup: Fixed input-field buttons not working (used a 'onchange' HTML attribute instead of 'onclick', whoops).