Cleaned up initial savefile implementation
Code-cleanup: Stored savefiles are now actually parsed when the site is loaded (previously it would simply use them directly without checking whether the savefile was even valid or not).
Code-cleanup: Cleaned up the documentation at the beginning of 'trackerCore.js' (and added mentions of a 'savefileless' savefile and savefiles storing settings which will both be implemented later).
Code-cleanup: Replaced all mentions of 'standAlone' with 'standalone' (yes, the capitalization is all that changed - yes, it was worth changing - no, I do not know why I wrote it the other way).