Finish cleaning up initial savefile implementation
Code-cleanup: Settings can now be validated (which is used when loading a savefile that stores settings).
Code-cleanup: When loading a savefile it now actually checks whether the game is valid or not.
Code-cleanup: Fix the validation of players (various things were rushed and just... not properly done at all).
Code-cleanup: Input-fields can properly verify whether a given color is valid or not.