Various input-field improvements
Technical changes: Input-fields can now have a custom ID (see 'id' in the input-field documentation under 'Attributes'). In addition input-field IDs ('fieldID') are now always in string-form; all functions that have accepted both number and string types now only accept strings.
Technical changes: Input-field labels no longer require the linked input-field to actually exist; whether it exists or not only matters once a user clicks on it. This is technically a bug-fix since the documentation always stated that this would be the case, it just kinda wasn't until now. Whoops.
Technical changes: 'inputfield_applyNewValue()' now supports a 'updateHost' argument. A bugfix relating to hosts not updating sometimes was also fixed by using this.
Technical changes: 'inputfield_setValue()' renamed the 'ignoreHost' argument to 'updateHost' (and also flipped the boolean value!!). This was done to be more accurate as to what it does; it doesn't ignore the host completely, it simply doesn't update the host (the host can't be ignored completely since some values needed are stored on the host instead of the input-field itself).