Characters can be changed again!
Fixed: Characters can now be changed again.

Technical-changes: The 'autoAddToForm' attribute is back but works differently now. It now has to be added to the form which will scan for descendants.
Technical-changes: 'formProperty' is no longer accepted for 'inputfield_setValue()'. Update the input-field directly instead of updating the form.
Code-cleanup: Restructued 'inputfield_setValue()'. This removes a lot of duplicate code that was already present in 'inputfield_applyNewValue()'.