HOST SUPPORT IS DONE
== Input-fields ==
Almost 5 months in the making!!

Technical-changes: Hosts now support all input-field types, not just checkboxes.
Technical-changes: Hosts can now be of any type (number, string, boolean, any object and functions), only exceptions are `undefined` and `null`.
Technical-changes: The 'belongsToHost' property on a FieldObject is now `null` instead of `false` when it doesn't belong to any host.
Technical-changes: The 'formChildren'  property on a FieldObject is now an object and includes 'fields' and 'hosts' arrays inside it to better differentiate them.
Technical-changes: Hosts now holds a 'defaultValue' property as well.
Technical-changes/Code-cleanup: Added a 'inputfield_verifyAddToForm()' to verify and apply the 'addToForm' attribute. This is now used instead of verifying them inside the 'FieldObject' and 'HostObject' constructors (the latter of which didn't even verify them).
Technical-changes: Rewrote the 'inputfield_setupHost()' function to fit these new changes. See the docs for details.
Technical-changes: Added a 'inputfield_getHostObject()' function to get the HostObject for a specific host.