input-fields suddenly got worse
== Input-fields ==
Technical-changes: Hosts now only supports strings, nothing else. This is important because objects are not allowed to save but how are you supposed to save which objects belong to a host when you're not allowed to save them? That's why it's only strings now. Could technically support numbers and booleans too but arguably it's bad practice to use these anyway and besides you can simply use a string that's just a number or a boolean.
Code-cleanup: Various descriptions and other minor fixes have been made. No functionality changed.