Input-fields~
== Input-field ==
Technical changes: Buttons now hold the value `null`.
Technical changes: Attributes are now properly synced with a host. At least on the field's creation, 'changeAttribute()' still doesn't update the host.
Technical changes: Fixed some input-fields not using the correct value after the field's creation. I honestly have no idea if this bug could even be encountered before this commit but regardless, it's fixed now! (issue was that 'FieldObject()' used 'specifics.value' instead of 'attributes.defaultValue')
Technical changes: Fixed some hosts not using the correct value after the host's creation. Same as above, no clue if this was an actual issue before. (issue was the 'defaultValue' attribute not being passed through)
Technical changes: Fixed checkboxes with a 'checkboxValue' attribute sometimes having an invalid 'displayedValue' property. Oh, you're not gonna believe this but I have no clue if it was an issue before!
Technical changes: Fixed the 'onchange' function not being called for hosts. Oh, oh! This one was actually an issue before!
Technical changes: Improved 'HostObject()' to include a lot of the improvements made to 'FieldObject()' in the last commit. It now takes different arguments, check the function's documentation for more info.
Technical changes: Related, the HostObject is now also structured differently to be similar to FieldObject.
Technical changes: 'verifyAddToForm()' was replaced by 'verifyFormAttributes()' which is essentially the same but better. It's also made with an upcoming 'addToForm' attribute rework in mind.
Code cleanup: Remove the 'file' type because it isn't implemented anyway (and won't be for a while).
Code cleanup: Changed the order in which attributes are verified in 'FieldObject()' to better reflect the order in which attributes are listed in the documentation. In addition there were also various improvements done which should have been a separate commit but I guess it's too late now, whoops.