INPUT-FIELDS ARE DONE!!
Nearly 8 months of work and it's finally done!
Well, there's still some improvements to be made here and there but it's finally at a useable spot where I can safely leave this project for a while without having to worry about not being able to get back into it!
With this, there's likely gonna be a couple more commits about this but then we should finally be back to regular MPO content!

== Input-fields ==
Technical changes: Hosts can now finally be removed and changed!! Simply use 'changeAttributes()' to change the 'host' attribute!
Technical changes: The 'data-labelforfieldid' was renamed to 'data-fieldlabelfor'.
Technical changes: Fixed an empty 'options' attribute breaking the site for input-fields of type 'radio'.
Code cleanup: Rewrote a huge chunk of the documentation.