Input-field cleanup duty!
== Input-field ==
Technical changes: The 'forms' attribute & the 'belongsToForm' property now consist of strings and not DOM elements. Should have happened earlier but didn't notice it, whoops.
Technical changes: Both 'getElement()' and 'getFieldObject()' now support a 'suppressError' argument to suppress the "couldn't be found" errors. This was also applied in all spots that already properly check and error-report the results of those two functions.
Technical changes: The 'options' attribute will now be properly verified when changing it through 'changeAttributes()'.
Code cleanup: It now checks whether a given element is a <input-field> element in a better, more robust way (previously it only checked whether the 'tagName' property was 'INPUT-FIELD' but now it checks whether a element was constructed by the 'InputFieldElement' class).
Code-cleanup: 'inputfield_getElement()' was rewritten to be both easier to read and (hopefully) faster.
Code cleanup: Made sure deleted input-fields are gracefully handled everywhere without logging warnings/errors to the console.