Code cleanup! Woo!!
Changed: Renamed 'savefileless' to 'settingsfile'.
This is a breaking internal change, if you've been using this website then press "Ctrl+Shift+I" to open DevTools, then click on "Storage" (if that doesn't exist then click on "Application"), select "Local Storage" on the left and select the only option inside it. Click on 'settingsfile' if it exists and press the Backspace/Delete key on your keyboard. Then doubleclick on 'savefileless' (found in the first column) and rename it to 'settingsfile'. Now reload the page.
Above explanation depends a lot on what browser you use but it should work on just about 99% of all current browsers (Chrome/Firefox and nearly everything that's based on it).

Code-cleanup: 'applySettings.js' has been cleaned up, mainly the documentations but some other things as well (could even call it "Fixed" instead of just "Cleaned up").
Internal changes: 'applySetting_applyAll()' renamed it's argument from 'setValues' to 'updateSetting'.
Internal changes: 'updatesTracker_updateCounter()' renamed it's argument from 'value' to 'amount' to reflect what it's actually called in the entire rest of the codebase. Whoops.
Internal changes: Renamed 'boot_basicStartup()' to 'boot_preStartup()'. Previous name was unclear and, to be quite honest, fucking stupid.