Minor counter changing features
Added: You can now click on a counter and enter the amount that counter should be. Just set "Action" to '=' and as long as the amount is '?' you can click on a counter to enter its amount. Can't be used on '+' or '-'.
Changed: If "Amount" is not set to 1, 5 or 10 then it just cycles to the next available number instead of starting over at 1.
Changed: 0 is now an accepted amount, though it can only be used for '=' (will be reset to 1 when you change the action to '+' or '-').

Code cleanup: Improved some docs. Not even a lot, just a bit.
Internal changes: renamed 'handleNavbar_updateAction()' to 'handleNavbar_updateActionDisplay()'
Internal changes: 'updatesTracker_setAmount()' can now handle NaN values instead of just pretending to. Whoops. (don't use '=== NaN', doesn't work.)
Internal changes: null is now an accepted 'amount'. Means "Handled by the interface", any calls to update a counter with null will be rejected.