Rewrote how mouse clicks are handled
Fixed: Added a failsafe in case the Action is something it shouldn't be. Don't think it fixed anything but maybe it will one day.
Code cleanup: Rewrote how mouse clicks are handled. There should be no noticeable difference.

Technical changes: Rewrote some of 'listeners_pointerDown()' and moved the list of switch cases into the 'listeners_pointerEventList' variable (setup in 'listeners_fillInPointerEventList()') to make it easier read & modify it and to make 'listeners.js' less dependant on other files (can be mostly copy-pasted to different projects now, explanations was added to the top of the file).