Internal stuff, woo!
Removed (?): Saving. Players & the selected savefile can't be saved. No clue how long this was broken or how many other saving-related things are also broken... but it's gonna have to wait a bit since Savefiles & Profiles as a whole is the next thing I'll be working on anyway. Not like you should be using the rewrite anyway at this point, lol.

Technical changes: Added fail-safes to various 'updatesTracker' functions in case you call them with an invalid player number (until now it only checked if it's a number).
Technical changes: 'trackerCore_status.savefiles.players' now starts at 1 instead of 0 thanks to being converted to an 'NonSequentialArray'. This is because the player number also starts at 1, so Player 1 is [1], Player 2 is [2]; you won't have to do a [players - 1] every time you update something related to a player.
Technical changes: Fixed 'NonSequentialArray.add()' not updating the '.length' property when specifying a slot.