Couple fixed caused by stupidity (my stupidity)
Fixed: Selected Savefile not being saved. It was a dumb mistake. A really dumb one. No, I do not want to explain.
Fixed: Characters not being saved. A bit more complicated but equally stupid. I made a custom solution to save players but it couldn't be converted to JSON properly (which the browser uses to save things permanently). Oh, and the custom solution didn't even work properly because it didn't even tell you all the players inside it. That was also a significant issue it had.

Technical changes: 'NonSequentialArray' now has a '.convertFrom()' property/function to convert an object (incl. arrays) to a NS-Array. Should be called like 'Array.isArray()', not on a NS-Array.
Technical changes: Fixed NS-Arrays not iterating. Yes, an entire feature that did not work properly despite me testing it. Please don't ask how that happened. But it was because I optimized it and properly forgot to test it properly afterwards instead of just checking its speed. Whoops.
Technical changes: The 'specifics.players' argument on 'trackerCore_Savefile' now supports Objects overall instead of just Arrays. Again, please don't ask how I missed this despite testing it.