Added MPS & SMPJ support, and some bugfixes
Oh wow, this is still active? Absolutely not, this commit is an outlier and should not be counted.

Added: MPS support to the tracker
Added: SMPJ support to the tracker AND the bonus star site!
Added: Bowser Space Bonus Star, related to the above two
Added: Pauline & Ninji as characters, related to SMPJ support

Changed: The active game is now highlighted with a yellow glow in the navbar.
Changed: Updated README to include a tutorial on how to add support for a new game, for future me and some other maniacs who wanna deal with this code.

Fixed: Ctrl & Shift hotkeys not working on Firefox. Blame Firefox not me, genuinely, why in the god damned fuck would they not send Ctrl & Shift events?!??!?! Btw, the "Ctrl" & "Shift" indicators that are seen in Chrome still aren't in Firefox because it is literally impossible to do. Not even kidding. Again, blame Firefox.
Fixed: Sick of hearing me complain about Firefox? Well you're in luck because I fixed a bug that caused the settings panel to not have an animation when opening it in Chrome when using the site as a local file. What do you mean ""Security Issue"", it's just a local file, what the hell?!

Code-cleanup: Made it so the site automatically adds new counters to the savefile instead of just crying about it and crashing the site. Only relevant when adding support for new games. Don't ask me how the site worked until now, I don't know.
Code-cleanup: Fixed a weird bug that hangs on boot when something else also broke at startup but that only started to happen in this commit (so shouldn't have been encountered until now).